+++ packages/twake-mui/src/lib/makeDarkOverrides.ts
@@ -1,19 +1,24 @@
 import { Theme, Components, alpha } from '@mui/material/styles'
 
+import { accordionOverrides } from './overrides/accordionOverrides'
 import { autocompleteOverrides } from './overrides/autocompleteOverrides'
 import { avatarOverrides } from './overrides/avatarOverrides'
 import { buttonOverrides } from './overrides/buttonOverrides'
+import { checkboxOverrides } from './overrides/checkboxOverrides'
+import { datePickerOverrides } from './overrides/datePickerOverrides'
 import {
   dialogOverrides,
   dialogTitleOverrides,
   dialogContentOverrides,
   dialogActionsOverrides
 } from './overrides/dialogOverrides'
+import { iconButtonOverrides } from './overrides/iconButtonOverrides'
 import { popoverOverrides } from './overrides/popoverOverrides'
 import {
   textFieldOverrides,
   outlinedInputOverrides,
-  inputBaseOverrides
+  inputBaseOverrides,
+  selectOverrides
 } from './overrides/textFieldOverrides'
 import {
   toggleButtonGroupOverrides,
@@ -21,11 +26,15 @@ import {
 } from './overrides/toggleButtonOverrides'
 
 export const makeDarkOverrides = (theme: Theme): Components => {
+  const datePickerOverridesResult = datePickerOverrides(theme)
+  const accordionOverridesResult = accordionOverrides(theme)
+
   return {
     MuiButton: buttonOverrides(theme),
     MuiTextField: textFieldOverrides(),
     MuiOutlinedInput: outlinedInputOverrides(),
     MuiInputBase: inputBaseOverrides(),
+    MuiSelect: selectOverrides(),
     MuiAutocomplete: autocompleteOverrides(),
     MuiDialog: dialogOverrides(),
     MuiDialogTitle: dialogTitleOverrides(),
@@ -35,6 +44,8 @@ export const makeDarkOverrides = (theme: Theme): Components => {
     MuiAvatar: avatarOverrides(theme),
     MuiToggleButtonGroup: toggleButtonGroupOverrides(),
     MuiToggleButton: toggleButtonOverrides(theme),
+    MuiIconButton: iconButtonOverrides(),
+    MuiCheckbox: checkboxOverrides(),
     MuiTypography: {
       styleOverrides: {
         root: {
@@ -43,6 +54,20 @@ export const makeDarkOverrides = (theme: Theme): Components => {
         caption: {
           // Typography caption overrides will be implemented later
         }
+      }
+    },
+    ...datePickerOverridesResult,
+    MuiAccordionSummary: accordionOverridesResult.MuiAccordionSummary,
+    MuiCssBaseline: {
+      styleOverrides: {
+        ...((datePickerOverridesResult.MuiCssBaseline?.styleOverrides as Record<
+          string,
+          unknown
+        >) || {}),
+        ...((accordionOverridesResult.MuiCssBaseline?.styleOverrides as Record<
+          string,
+          unknown
+        >) || {})
       }
     }
   }