+++ packages/twake-mui/src/lib/overrides/checkboxOverrides.ts
@@ -1,0 +1,13 @@
+import { Components } from '@mui/material/styles'
+
+export const checkboxOverrides = (): Components['MuiCheckbox'] => {
+  return {
+    styleOverrides: {
+      root: {
+        '&.MuiCheckbox-sizeSmall': {
+          padding: '6px'
+        }
+      }
+    }
+  }
+}